chore: refactor steps out of clack utils
