+++ src/react-native/react-native-wizard.ts
@@ -4,7 +4,6 @@ import {
   abort,
   askForAIConsent,
   confirmContinueIfNoOrDirtyGitRepo,
-  createPRFromNewBranch,
   ensurePackageIsInstalled,
   getOrAskForProjectData,
   getPackageDotJson,
@@ -12,7 +11,6 @@ import {
   installPackage,
   isUsingTypeScript,
   printWelcome,
-  runPrettierIfInstalled,
 } from '../utils/clack-utils';
 import { getPackageVersion, hasPackageInstalled } from '../utils/package-json';
 import clack from '../utils/clack';
@@ -26,9 +24,9 @@ import {
 } from '../utils/file-utils';
 import type { WizardOptions } from '../utils/types';
 import { askForCloudRegion } from '../utils/clack-utils';
-import { addEditorRules } from '../utils/rules/add-editor-rules';
+import { addEditorRulesStep, runPrettierStep } from '../steps';
 import { EXPO } from '../utils/package-manager';
-import { getOutroMessage, getPRDescription } from '../lib/messages';
+import { getOutroMessage } from '../lib/messages';
 
 export async function runReactNativeWizard(
   options: WizardOptions,
@@ -143,12 +141,12 @@ export async function runReactNativeWizard(
     cloudRegion,
   });
 
-  await runPrettierIfInstalled({
+  await runPrettierStep({
     installDir: options.installDir,
     integration: Integration.reactNative,
   });
 
-  const addedEditorRules = await addEditorRules({
+  const addedEditorRules = await addEditorRulesStep({
     installDir: options.installDir,
     rulesName: 'react-native-rules.md',
     integration: Integration.reactNative,
@@ -159,24 +157,12 @@ export async function runReactNativeWizard(
     installDir: options.installDir,
   });
 
-  const prDescription = getPRDescription({
-    integration: Integration.reactNative,
-    addedEditorRules,
-  });
-
-  const prUrl = await createPRFromNewBranch({
-    installDir: options.installDir,
-    integration: Integration.reactNative,
-    body: prDescription,
-  });
-
   const outroMessage = getOutroMessage({
     options,
     integration: Integration.reactNative,
     cloudRegion,
     addedEditorRules,
     packageManager: packageManagerForOutro,
-    prUrl,
   });
 
   clack.outro(outroMessage);
