+++ src/react/react-wizard.ts
@@ -4,7 +4,6 @@ import {
   abort,
   askForAIConsent,
   confirmContinueIfNoOrDirtyGitRepo,
-  createPRFromNewBranch,
   ensurePackageIsInstalled,
   getOrAskForProjectData,
   getPackageDotJson,
@@ -12,17 +11,13 @@ import {
   installPackage,
   isUsingTypeScript,
   printWelcome,
-  runPrettierIfInstalled,
 } from '../utils/clack-utils';
 import { getPackageVersion, hasPackageInstalled } from '../utils/package-json';
 import clack from '../utils/clack';
 import { Integration } from '../lib/constants';
 import { getReactDocumentation } from './docs';
 import { analytics } from '../utils/analytics';
-import {
-  addOrUpdateEnvironmentVariables,
-  detectEnvVarPrefix,
-} from '../utils/environment';
+import { detectEnvVarPrefix } from '../utils/environment';
 import {
   generateFileChangesForIntegration,
   getFilesToChange,
@@ -30,8 +25,12 @@ import {
 } from '../utils/file-utils';
 import type { WizardOptions } from '../utils/types';
 import { askForCloudRegion } from '../utils/clack-utils';
-import { addEditorRules } from '../utils/rules/add-editor-rules';
-import { getOutroMessage, getPRDescription } from '../lib/messages';
+import { getOutroMessage } from '../lib/messages';
+import {
+  addEditorRulesStep,
+  addOrUpdateEnvironmentVariablesStep,
+  runPrettierStep,
+} from '../steps';
 
 export async function runReactWizard(options: WizardOptions): Promise<void> {
   printWelcome({
@@ -114,7 +113,7 @@ export async function runReactWizard(options: WizardOptions): Promise<void> {
     cloudRegion,
   });
 
-  await addOrUpdateEnvironmentVariables({
+  await addOrUpdateEnvironmentVariablesStep({
     variables: {
       [envVarPrefix + 'POSTHOG_KEY']: projectApiKey,
       [envVarPrefix + 'POSTHOG_HOST']: host,
@@ -126,27 +125,16 @@ export async function runReactWizard(options: WizardOptions): Promise<void> {
   const packageManagerForOutro =
     packageManagerFromInstallStep ?? (await getPackageManager(options));
 
-  await runPrettierIfInstalled({
+  await runPrettierStep({
     installDir: options.installDir,
     integration: Integration.react,
   });
 
-  const addedEditorRules = await addEditorRules({
+  const addedEditorRules = await addEditorRulesStep({
     installDir: options.installDir,
     rulesName: 'react-rules.md',
     integration: Integration.react,
     default: options.default,
-  });
-
-  const prDescription = getPRDescription({
-    integration: Integration.react,
-    addedEditorRules,
-  });
-
-  const prUrl = await createPRFromNewBranch({
-    installDir: options.installDir,
-    integration: Integration.react,
-    body: prDescription,
   });
 
   const outroMessage = getOutroMessage({
@@ -155,7 +143,6 @@ export async function runReactWizard(options: WizardOptions): Promise<void> {
     cloudRegion,
     addedEditorRules,
     packageManager: packageManagerForOutro,
-    prUrl,
   });
 
   clack.outro(outroMessage);
