+++ src/steps/add-editor-rules.ts
@@ -1,51 +1,51 @@
 import * as fs from 'fs';
 import chalk from 'chalk';
 import path from 'path';
-import { Integration } from '../../lib/constants';
-import { analytics } from '../analytics';
-import clack from '../clack';
-import { traceStep } from '../../telemetry';
-import { abortIfCancelled } from '../clack-utils';
+import { Integration } from '../lib/constants';
+import { analytics } from '../utils/analytics';
+import clack from '../utils/clack';
+import { traceStep } from '../telemetry';
+import { abortIfCancelled } from '../utils/clack-utils';
 
-type AddEditorRulesOptions = {
+type AddEditorRulesStepOptions = {
   installDir: string;
   rulesName: string;
   integration: Integration;
   default?: boolean;
 };
 
-export const addEditorRules = async ({
+export const addEditorRulesStep = async ({
   installDir,
   rulesName,
   integration,
   default: defaultAddEditorRules,
-}: AddEditorRulesOptions): Promise<boolean> => {
+}: AddEditorRulesStepOptions): Promise<boolean> => {
   // Add rules file if in Cursor environment
   if (process.env.CURSOR_TRACE_ID) {
-    const addEditorRules: boolean = defaultAddEditorRules
-      ? true
-      : await abortIfCancelled(
-          clack.select({
-            message:
-              'Would you like to have PostHog added to your Cursor rules?',
-            options: [
-              {
-                label: 'Yes, please!',
-                value: true,
-              },
-              {
-                label: 'No, thanks',
-                value: false,
-              },
-            ],
-          }),
-        );
-
-    if (!addEditorRules) {
-      return false;
-    }
-
     return traceStep('add-editor-rules', async () => {
+      const addEditorRules: boolean = defaultAddEditorRules
+        ? true
+        : await abortIfCancelled(
+            clack.select({
+              message:
+                'Would you like to have PostHog added to your Cursor rules?',
+              options: [
+                {
+                  label: 'Yes, please!',
+                  value: true,
+                },
+                {
+                  label: 'No, thanks',
+                  value: false,
+                },
+              ],
+            }),
+          );
+
+      if (!addEditorRules) {
+        return false;
+      }
+
       const docsDir = path.join(installDir, '.cursor', 'rules');
 
       await fs.promises.mkdir(docsDir, { recursive: true });
@@ -76,7 +76,9 @@ export const addEditorRules = async ({
         integration,
       });
 
-      clack.log.info(`Added Cursor rules to ${chalk.bold.cyan(docsDir)}`);
+      clack.log.info(
+        `Added Cursor rules to ${chalk.bold.cyan(`.cursor/rules`)}`,
+      );
 
       return true;
     });
